+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # ChangeLog
 
+- support Time column DuckDB#Result#chunk_each
 - add DuckDB::Interval#eql?
 - drop Ruby 2.7.x.
 
@@ -40,7 +41,7 @@
 - deprecation warning when DuckDB::Result.each calling with `DuckDB::Result.use_chunk_each` is false.
   The `each` behavior will be same as `DuckDB::Result.chunk_each` in the future.
   set `DuckDB::Result.use_chunk_each = true` to suppress the warning.
-- DuckDB::Result#chunck_each returns DuckDB::Interval class when the column type is INTERVAL.
+- DuckDB::Result#chunk_each returns DuckDB::Interval class when the column type is INTERVAL.
 
 # 0.8.1.3
 - Fix BigDecimal conversion.