+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # ChangeLog
 
+# 0.9.2.1
 - support Time column in `DuckDB#Result#chunk_each`.
 - add `DuckDB::Interval#eql?`.